% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myconstr.R
\name{myconstr}
\alias{myconstr}
\title{My constructor function}
\usage{
myconstr(x, y, alpha)
}
\arguments{
\item{x}{X data}

\item{y}{Y data}

\item{alpha}{alpha level}
}
\value{
a list of data, alpha, confidence interval
}
\description{
My constructor function
}
\examples{
\dontrun{myconstr(x = x, y = y, alpha = alpha)}
}
