% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Rttest.R
\name{plot.Rttest}
\alias{plot.Rttest}
\title{plot method for Rttest}
\usage{
\method{plot}{Rttest}(x, ...)
}
\arguments{
\item{...}{additional variables}

\item{z}{output from Rttest constructor}
}
\value{
side by side boxplot of data
}
\description{
plot method for Rttest
}
\examples{
\dontrun{plot(out)}
}
