% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Rttest.R
\name{print.Rttest}
\alias{print.Rttest}
\title{print method for Rttest}
\usage{
\method{print}{Rttest}(x, ...)
}
\arguments{
\item{x}{dataframe to print}

\item{...}{additional variables}
}
\value{
table of dataframe using kable
}
\description{
print method for Rttest
}
\examples{
\dontrun{print(out)}
}
